test: split PayPal dispute and identity recovery Postgres tests
